Restrict "id" parameter type to int or string
Changes were made to the Resizer and ResizerEventHandlers class, "id" parameter in the constructor function and some methods now only accept int or string, removing the capability to accept array. This simplification streamlines the implemented methods and ensures data consistency; while also handling the conversion of type, before passing to other methods. Apart from this, made a switch in the default imageResizerClass, from Grafika to Spatie for increased efficiency.